+++ lib/screens/home_tabs/new_booking.dart
@@ -9,9 +9,20 @@ class _HallsView extends StatelessWidget {
       create: (context) => HallsBloc()..add(const HallsEvent.update()),
       child: BlocBuilder<HallsBloc, HallsState>(
         builder: (context, state) => state.when(
-          success: (halls) => Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: halls.map((hall) => _HallCard(hall: hall)).toList(),
+          success: (halls, hallsMobile) => SingleChildScrollView(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: halls
+                  .map(
+                    (hall) => _HallCard(
+                      hall: hall,
+                      hallMobile: hallsMobile.firstWhere(
+                        (element) => element.name == hall.hname,
+                      ),
+                    ),
+                  )
+                  .toList(),
+            ),
           ),
           loading: () => const LoadingWidget(),
           error: (e) => CenteredText(
@@ -25,11 +36,15 @@ class _HallsView extends StatelessWidget {
 
 class _HallCard extends StatelessWidget {
   final Hall hall;
-  const _HallCard({Key? key, required this.hall}) : super(key: key);
+  final HallMobile hallMobile;
+  const _HallCard({Key? key, required this.hall, required this.hallMobile})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Card(
+      elevation: 10.0,
+      margin: const EdgeInsets.symmetric(vertical: 12.0, horizontal: 16.0),
       child: InkWell(
         onTap: () => Navigator.of(context).push(
           MaterialPageRoute(
@@ -40,8 +55,31 @@ class _HallCard extends StatelessWidget {
           ),
         ),
         child: Padding(
-          padding: const EdgeInsets.all(32.0),
-          child: Text(hall.building),
+          padding: const EdgeInsets.all(16.0),
+          child: SizedBox(
+            height: 70,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  hallMobile.name,
+                  style: const TextStyle(
+                    fontSize: 18.0,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    const Icon(Icons.location_pin, size: 16),
+                    Flexible(child: Text(hallMobile.location)),
+                  ],
+                ),
+              ],
+            ),
+          ),
         ),
       ),
     );