+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Redesign of the bookings page
+- Redesign of the halls page
 
 ## [0.1.8] - 2022-07-11
 - Add settings page